chore: move task group and task definition to corral
