+++ src/corral/base.py
@@ -1,13 +1,9 @@
-from __future__ import annotations
-
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any
-
-if TYPE_CHECKING:
-    from collections.abc import Callable
+from collections.abc import Callable
 from dataclasses import dataclass, field
 from datetime import datetime, timezone
 from enum import Enum, StrEnum
+from typing import Any
 
 from pydantic import BaseModel
 
@@ -293,3 +289,105 @@ Example tool call format:
         self.state.is_completed = True
         self.state.end_time = datetime.now(tz=timezone.utc)
         return score
+
+
+### Classes related to subtasks
+
+
+@dataclass
+class TaskDefinition:
+    """Definition of a task with its requirements and scoring"""
+
+    name: str
+    description: str
+    tools: list[str]
+    scoring_fn: Callable[[dict | str], float]
+    submission_format: dict[str, str]
+    # Either use output from another task or custom input
+    input_from_tasks: list[str] = field(
+        default_factory=list
+    )  # input required from some of the previous tasks in the group
+    initial_input: dict[str, Any] = field(
+        default_factory=dict
+    )  # initial input for the task, if required
+
+    # Helper method to check if task has dependencies
+    def has_dependencies(self) -> bool:
+        return len(self.input_from_tasks) > 0
+
+
+@dataclass
+class TaskGroup:
+    """Container for related tasks"""
+
+    group_id: str
+    tasks: dict[str, TaskDefinition]
+    results: dict[str, Any] = field(default_factory=dict)
+    scores: dict[str, float] = field(default_factory=dict)
+
+    def get_task_input(self, task_id: str) -> dict[str, Any]:
+        """Get input for a task either from other tasks or initial input"""
+        task = self.tasks.get(task_id)
+        if not task:
+            return {}
+
+        # Start with the initial input
+        combined_input = task.initial_input.copy() if task.initial_input else {}
+
+        # Add inputs from dependent tasks
+        for dep_task_id in task.input_from_tasks:
+            if dep_task_id in self.results:
+                # Add the dependent task's result to the input
+                dep_result = self.results[dep_task_id]
+                if isinstance(dep_result, dict) and "answer" in dep_result:
+                    # Extract the relevant part of the result
+                    combined_input[f"result_from_{dep_task_id}"] = dep_result["answer"]
+                else:
+                    combined_input[f"result_from_{dep_task_id}"] = dep_result
+
+        return combined_input
+
+    def store_result(self, task_id: str, result: dict[str, Any], score: float) -> None:
+        """Store task result and score"""
+        self.results[task_id] = result
+        self.scores[task_id] = score
+
+    def get_task_dependencies(self) -> dict[str, list[str]]:
+        """Get dictionary of task dependencies"""
+        return {task_id: task.input_from_tasks for task_id, task in self.tasks.items()}
+
+    def get_ordered_tasks(self) -> list[str]:
+        """Return tasks in dependency order"""
+        # Simple topological sort
+        dependencies = self.get_task_dependencies()
+        visited = set()
+        ordered = []
+
+        def visit(task_id):
+            if task_id in visited:
+                return
+            visited.add(task_id)
+            for dep in dependencies.get(task_id, []):
+                visit(dep)
+            ordered.append(task_id)
+
+        for task_id in self.tasks:
+            visit(task_id)
+
+        return ordered
+
+    def check_dependencies_satisfied(self, task_id: str) -> bool:
+        """Check if all dependencies for a task are satisfied"""
+        task = self.tasks.get(task_id)
+        if not task:
+            return False
+
+        return all(
+            dep_task_id in self.results and self.results[dep_task_id] is not None
+            for dep_task_id in task.input_from_tasks
+        )
+
+
+@dataclass
+class TaskGroups:
+    pass
